+++ scalyr_agent/tests/monitors_manager_test.py
@@ -20,7 +20,6 @@ from __future__ import absolute_import
 
 __author__ = "czerwin@scalyr.com"
 
-
 import re
 
 import scalyr_agent.util as scalyr_util
@@ -39,7 +38,6 @@ class MonitorsManagerTest(ScalyrTestCase):
             [],
         )
         self.assertEquals(len(test_manager.monitors), 1)
-        self.assertEquals(test_manager.monitors[0].monitor_name, "test_monitor()")
 
     def test_multiple_modules_and_platform_monitors(self):
         test_manager, _ = ScalyrTestUtils.create_test_monitors_manager(
@@ -293,7 +291,7 @@ class MonitorsManagerTest(ScalyrTestCase):
         test_manager.set_user_agent_augment_callback(augment_user_agent_2)
 
         self.assertTrue(
-            fragment_polls.sleep_until_count_or_maxwait(20, poll_interval, 2)
+            fragment_polls.sleep_until_count_or_maxwait(20, poll_interval, 2.5)
         )
         self.assertEquals(counter["callback_invocations"], 10)